+++ benchmark/vector/vector_initializing/float32_from_list.dart
@@ -1,22 +1,27 @@
-// Approx. 2.5 sec (MacBook Air mid 2017)
+// Approx. 6.5 sec (MacBook Air mid 2017), Dart VM version: 2.5.0
 
 import 'package:benchmark_harness/benchmark_harness.dart';
+import 'package:ml_linalg/linalg.dart';
 import 'package:ml_linalg/src/vector/float32/float32_vector.dart';
 
 const amountOfElements = 10000000;
 
 class VectorFromListBenchmark extends BenchmarkBase {
-  const VectorFromListBenchmark()
+  VectorFromListBenchmark()
       : super('Vector initialization (fromList), '
       '$amountOfElements elements');
 
+  final _source = Vector
+      .randomFilled(amountOfElements, min: -10000, max: 10000)
+      .toList();
+
   static void main() {
-    const VectorFromListBenchmark().report();
+    VectorFromListBenchmark().report();
   }
 
   @override
   void run() {
-    Float32Vector.fromList(List<double>.filled(amountOfElements, 1.0));
+    Float32Vector.fromList(_source);
   }
 }