+++ CHANGELOG.md
@@ -5,7 +5,10 @@
     - `pick` method removed
     - `submatrix` method replaced with `sample` method
     - `[]` operator returns `Vector` from now
-- `Vector`: `randomFilled` constructor api changed, `min` and `max` args are of `num` type now
+- `Vector`: 
+    - `randomFilled` constructor api changed, `min` and `max` args are of `num` type now
+    - `fromList` constructor api changed, from now as a source expected `List<num>` instead of `List<doublet>`
+    - `filled` constructor api changed, from now sampling value is a `num` instead of `duble`
 - Dart sdk constraint returned to 2.2.0
 
 ## 11.0.0